Remove deprecated authentication controllers and update package-lock.json
Removed deprecated authentication-related controllers as a part of remodeling the authentication system within the application. The package-lock.json file was also updated as a part of a routine update to keep npm dependencies up to date. Removed files include controllers such as ConfirmPasswordController, ForgotPasswordController, and LoginController from both the auth and auth2 paths, among others. This was done in preparation for introducing new and updated authentication controllers to replace the deleted ones in order to improve security and efficiency.