Update composer dependencies and refactor user registration process
Updated composer.lock to include a new version of PHP and the "spatie/data-transfer-object" library.
The User.php model was updated to change '$fillable' attribute to '$guarded'.
The registration process previously located in RegisterController.php was moved to a new service class (RegisterAction.php) to make code more modular and easy to maintain.
This also led to creating several new files for the registration data transfer object (DTO) and related request handling.
ManagerSeeder.php was added for seeding manager permissions. The minimum PHP version was updated in composer.json.