"Refactor database migrations and add auth register service
Several database migrations have been deleted. This is part of a larger effort to clean up the database schema and eliminate unused tables and columns.

In addition, the 'app/Services/Auth' directory has several new files and changes related to the implementation of a new authentication service. This includes a new service provider (ActionServiceProvider), a DTO factory (RegisterDtoFactory), and interfaces and classes for a register action (RegisterAction and Register). The updates in these files are part of a new registration feature where the user's name, email, and password are recorded.

Furthermore, a new dependency 'spatie/laravel-permission' is added to handle user permissions and roles in the application."